requirements.txt
duckdb==1.2.2
numpy==2.2.6
pandas==2.2.3
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2

# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5

# Google AI dependencies
google-generativeai>=0.3.0

# Database dependencies
duckdb>=0.9.0
pandas>=2.0.0

# Vector store dependencies
qdrant-client>=1.6.0

# Environment management
python-dotenv>=1.0.0